+++ src/app/shared/transaction-result.component.ts
@@ -3,7 +3,7 @@ import * as moment from 'moment';
 
 interface TransactionData {
   seller: number;
-  value: string;
+  value: number;
   purchase_time: string;
 }