+++ src/app/dashboard/dashboard.routing.ts
@@ -12,6 +12,8 @@ import { AccountEditComponent } from './account-edit.component';
 import { AddDataComponent } from './add-data.component';
 import { FeedbackComponent } from './feedback.component';
 import { TransactionLogComponent } from './transaction-log.component';
+import { PayrollLogComponent } from './payroll-log.component';
+import { MapComponent } from './map.component';
 
 // Using child path to allow for FullLayout theming
 const routes: Routes = [
@@ -47,6 +49,18 @@ const routes: Routes = [
         path: 'transaction-log',
         component: TransactionLogComponent,
         data: { title: 'Transaction Log' },
+      },
+      {
+        path: 'map',
+        component: MapComponent,
+        data: { title: 'Supplier Map' },
+        canActivate: [OrgGuard],
+      },
+      {
+        path: 'payroll-log',
+        component: PayrollLogComponent,
+        data: { title: 'Payroll Log' },
+        canActivate: [OrgGuard],
       },
       {
         path: 'feedback',