+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 # Next Release
 
+# v0.0.2
+
+* Supplier map added to organisation view
+* Fix various linter errors
+* Payroll Log Data added
+* Fixed register box view on login page
+* Made the form reset properly if adding a new one
+* Added better error messages on register and account edit
+
 # v0.0.1
 
 * Added Travis Testing